% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookup}
\alias{lookup}
\title{lookup tables}
\format{A data frame with columns:
\describe{
 \item{lookup}{lookup table name}
 \item{code}{code}
 \item{description}{description of the code}

}}
\source{
this table includes the full description of the codes used in the wolfdiet table.
}
\usage{
lookup
}
\description{
lookup tables
}
\examples{
\dontrun{
 lookup
}
}
\keyword{datasets}
