% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{wolf}
\alias{wolf}
\title{wolf}
\usage{
wolf(tax = TRUE)
}
\arguments{
\item{tax}{species represented using scientific names (TRUE) or common names (FALSE)}
}
\description{
wolf
}
\examples{
wolf()

}
