% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wolfdiet}
\alias{wolfdiet}
\title{wolfdiet view}
\format{An object of class \code{data.frame} with 515 rows and 94 columns.}
\source{
this table compiles wolf diet from around the world - currently Eurasia.
}
\usage{
wolfdiet
}
\description{
wolfdiet view
}
\examples{
\dontrun{
 wolfdiet
}
}
\keyword{datasets}
